refactor(): extract TopNav & SideNav
