+++ src/components/SideNav.vue
@@ -1,11 +1,4 @@
 <script setup lang="ts">
-import MenuCard from './Cashier/MenuCard.vue'
-import SalesCard from './Cashier/SalesCard.vue'
-
-defineProps<{ msg: string }>()
-
-const topNavLeft = ["Home", "Name"]
-const topNavRight = ["Logout", "Kuntau"]
 const sideNav = [
   {
     name: 'Dashboard',
@@ -50,9 +43,7 @@ const sideNav = [
     active: false
   }
 ]
-
 </script>
-
 <template>
   <div class="sidebar fixed px-2 pl-5 py-3">
     <div class="store text-blue-800 font-medium px-1 py-2 mb-3">
@@ -78,49 +69,7 @@ const sideNav = [
       </ul>
     </nav>
   </div>
-
-  <section class="main container box-border relative w-full">
-
-    <nav class="top-nav sticky inset-0 box-border inline-flex flex-grow w-full h-auto py-1 overflow-hidden">
-      <div class="top-nav-left flex flex-grow items-start">
-        <ul class="flex self-center">
-          <li class="text-gray-700" v-for="nav in topNavLeft">{{ nav }}</li>
-        </ul>
-        <div class="inline-flex self-center">
-          <input class="search ml-2 px-4 h-8 content-center shadow rounded-l" type="text" placeholder='Press / to focus search'>
-          <button class="btn inline-block text-white text-core bg-green-500 hover:bg-green-600 h-8 px-2 -ml-1 rounded-r shadow"><i class="fas fa-search text-xl align-text-bottom leading-none"></i></button>
-        </div>
-      </div>
-      <div class="top-nav-right">
-        <ul class="flex justify-end items-center">
-          <li class="inline-block" v-for="nav in topNavRight">
-            <a class="text-gray-700" href="#">
-              <i class="fad fa-sign-in"></i> {{ nav }}
-            </a>
-          </li>
-          <li class="inline-block"><a href="#"><img class="h-8 rounded-full shadow" src="https://gravatar.com/avatar/d69aae377ffa1f6fd90935a888933070" alt=""></a></li>
-        </ul>
-      </div>
-    </nav>
-
-    <content class="box-border my-10 flex">
-      <MenuCard />
-      <SalesCard />
-    </content>
-  </section>
 </template>
-
 <style scoped>
 
-#label {
-  margin: 0 0.5em;
-  font-weight: bold;
-}
-
-#code {
-  background-color: #eee;
-  padding: 2px 4px;
-  border-radius: 4px;
-  color: #304455;
-}
 </style>
